Do not gracefully drain the proxy on SIGTERM
Per the discussion in linkerd/linkerd2#10379, the proxy's shutdown
behavior is overly aggressive and does not properly honor a pod's
`terminationGracePeriodSeconds` configuration: while an application
container may continue to run with a server for some period after the
pod's delete process begins, the proxy will not permit any new
connections in or out of the pod. This may also interfere with
administrative probes during the shutdown process.

To address this, the proxy is updated to only change its readiness
probe state when it receives a SIGTERM. The proxy continues to run--
without closing any connections or rejecting any work--until it is
terminated (i.e., by kubelet) with a SIGKILL.

It is expected that, in most cases, clients will voluntarily move
traffic as they process discovery updates. We need not do anything to
interfere with the application's graceful termination behavior.

The proxy continues its former graceful shutdown behavior when shutdown
is initiated via the admin server `/shutdown` endpoint (e.g., by
linkerd-await).