stack: Make `failfast::Gate` general purpose (#2279)
`failfast::Gate` is a stack middleware that supports changing its
readiness based on some external signal (the failfast state changing).
Nothing about the gate's behavior is specific to failfast, though; and
this type of gating is general useful on its own.

To setup further reuse, this change moves `stack::failfast::gate` to
`stack::gate`. The gate is now tested independently from failfast.
Failfast continues to maintain an (optional) `gate::Tx` so that it can
control a gate.