integration: add test policy controller
Currently, the proxy integration tests run without a policy controller,
and instead use environment variables to configure default inbound
server policies. However, we do not intend to add a similar environment
based defaulting mechanism for outbound client policies. Therefore, when
the proxy starts consuming the outbound client policy API, we will no
longer be able to run the integration tests without a policy controller.

This branch adds an implementation of a mock policy controller for the
integration tests. This was factored out from PR #2265, which implements
outbound policy discovery and therefore requires a policy controller.

Note that the mock policy controller added in #2265 serves both the
inbound and outbound APIs, while this PR implements only the inbound
API. This is because the a version of `linkerd2-proxy-api` that includes
the outbound policy API has not yet been released, and I wanted to be
able to merge this branch without waiting for a proxy API release.
However, the mock policy controller implementation here was taken
directly from #2265 with the outbound half removed. Some aspects of the
implementation may seem somewhat more complex than necessary to
implement only the inbound policy API, but I thought it was nicer to
have a smaller diff when the outbound half is added in the near future.

One small complexity here is that the inbound proxy will use default
policies when it has not yet resolved. This means there is a potential
for racy behavior in tests when a request is sent to an inbound proxy,
if it has not yet started a watch on the policy for a port (or is in the
process of starting that watch). This can result in flaky test failures:
for example, in metrics tests, the name of the policy can change between
requests if one request is handled with the default and the next request
is handled with one resolved from the mock policy controller, resulting
in two sets of labels rather than one. To avoid the test proxy flapping
between the default policy and one resolved from the mock policy
controller, I've also changed the proxy test harness so that any
configured inbound ports in the test are added to
`LINKERD2_PROXY_INBOUND_PORTS`, so that the proxy will eagerly resolve
them. The test harness now waits for the mock policy controller to tell
it that the proxy has started watches on those ports before actually
running the test code.