client-policy: add protobuf conversion (#2289)
This branch adds functions to the `linkerd-proxy-client-policy` crate
for converting the `linkerd2-proxy-api` outbound policy protobuf types
into the proxy's internal representations. These are not used in this
branch, but will be used in subsequent changes that actually wire up
outbound policy discovery.

This change was factored out from #2265, so that it can be reviewed
independently of the rest of the client policy discovery plumbing.

Note that this branch changes the `linkerd2-proxy-api` dependency to a
Git dep on the outbound policy branch (linkerd/linkerd2-proxy-api#165),
as the new APIs have not yet been published.