outbound: Report concrete authorities for policies (#2313)
The new policy router currently reports a numeric authority when using
policy routes. In some cases, we have a named concrete address for the
load balancer. In the vast majority of cases, this address is the same
as the logical service's.

For now, let's use concrete addresses for telemetry. This will help
minimize regressions while we figure how to move telemetry forward.