test: add mock client policy resolver (#2314)
Currently, `linkerd-app-outbound`'s stack tests construct a discovery
service for `push_discover` which only discovers service profiles, and
does not resolve client policies. This works fine currently, as
`Discovery` can be constructed with only a profile receiver, as the
client policy receiver is optional.

However, when we add client policy discovery to the gateway, we will be
able to make the policy receiver no longer optional, as a client policy
will always be discovered. In that case, these tests will no longer
compile. Therefore, this branch adds an implementation of a mock client
policy resolver in `linkerd-test`, and updates these stack tests to
actually use it.

This was factored out of the change that actually adds policy discovery
to the gateway, and is a prerequisite for that change.