outbound: fix `Balance::Dispatch` "authority" labels
PR #2313 changed client policies with the load balancer dispatch type to
report the load balancer's destination address as the "authority" label,
rather than the numeric authority the policy was discovered for.
However, this change was accidentally undone when merging PR #2260,
which moved the code where the authority label is generated to a
different file.

This PR changes it back, so that the discovered concrete destination
address should still be reported as the "authority" metrics label.