gate: Support admitting a limited number of requests (#2342)
To implement circuit breakers, we want to use the gate to control
whether or not a service is available to callers. When a service is
coming out of a failed state, we may want to admit a limited number of
requests through before going back to an unfettered state.

To support this, this change updates the `Gate` service to be backed by
a `tokio::sync::watch` channel so that each side of the gate channel can
communicate state updates.

`gate::Tx::closed` is renamed to `gate::Tx::lost` to avoid ambiguity.