add `NewClassifyGate` and `BroadcastClassification` middleware
This branch introduces middleware for opening and closing a `Gate`
middleware based on HTTP response classification. The
`BroadcastClassification` service applies an HTTP response classifier to
each request, and sends the classification over a channel. The
`NewClassifyGate` middleware constructs stacks containing
`BroadcastClassification` and `Gate` middleware. In conjunction with a
_failure accrual policy_, these middleware allow response classification
results to determine whether a service stack is available, implementing
_request-level HTTP circuit breaking_.

Failure accrual policies are determined by the `ExtractParam`
implementation provided to the `NewClassifyGate` layer. This type must
provide a `classify::gate::Params` type containing a `mpsc::Sender` for
the response classification channel, and a `gate::Rx` that controls the
opening and closing of the `Gate` middleware. This means that the
`ExtractParam` implementation can spawn a task implementing an arbitrary
policy for opening and closing the gate based on classifications
received over the `mpsc::Receiver` end of the classification channel,
based on the target type the stack is constructed for.

This branch does not implement any failure accrual policies, nor does it
integrate the `NewClassifyGate` middleware with existing proxy stacks.
It simply adds these middleware. Subsequent branches will add the
circuit-breaking middleware to the proxy stack, implement failure accrual
policies, and configure failure accrual from individual stack targets.

This change was factored out from PR #2334 in order to merge
independently.