client-policy: `RouteBackend`s use the backend's metadata (#2351)
Due to an oversight in the code for converting the proxy API
`OutboundPolicy` protobuf types into the proxy's internal
representations, `RouteBackend`s are currently constructed with the
_route_'s metadata, rather than the metadata for that individual
backend. This is incorrect, so this branch fixes that.

Because the functions for converting the protobuf types to
`RouteBackend`s and `RouteDistribution`s no longer need to have the
route's `Arc<Meta>` passed in, some of these functions have been changed
to `TryFrom` implementations.