client-policy: accept backends with empty `dispatcher` fields (#2352)
When a `backendRef` in an HTTPRoute references a valid Service name
which doesn't currently exist in the cluster, the policy controller
constructs a synthesized backend with a `failureInjector` filter, so
that traffic sent to the invalid backend will fail with a descriptive
error message, and any other backends in the route will still work
normally. Because the backend doesn't exist, there is no reasonable
value for the policy controller to set as the backend's `dispatcher`
field in the OutboundPolicy proto.

Unfortunately, however, the proxy currently rejects any OutboundPolicy
which has a backend without a populated `dispatcher` field. This means
that when an OutboundPolicy has one backend that has no `dispatcher`,
the proxy will synthesize an invalid policy that fails all traffic,
rather than one that only fails traffic sent to that backend.

This branch changes the proxy so that backends without a `dispatcher`
are no longer rejected by the client policy protobuf conversion code.
Instead, a new `Dispatcher::Fail` variant is used to represent a backend
which always fails requests with a configured error message, so that a
functional `ClientPolicy` can still be constructed when one or more
backends lack a dispatcher. When constructing a stack for a `Concrete`
target with a `Fail` dispatcher, the proxy builds a service that fails
all requests. In practice, these errors will not be observed when the
policy controller returns a backend for an invalid Service name, as the
`failureInjector` filter's error should be produced first, but
constructing a service is necessary for the stack to function correctly.

Closes linkerd/linkerd2#10620