gate: fix `Gate` trying to acquire multiple semaphore permits
The `Gate` middleware's `poll_ready` implementation attempts to acquire
a semaphore permit when the gate is in a limited state, and then polls
the inner service for readiness. If the inner service is ready, the
semaphore permit is stored in the gate to be consumed in `call`.
However, this implementation does not check if the gate has *already*
acquired a sempahore permit and stored it in the gate. In that case, a
subsequent call to `poll_ready` after the gate has returned
`Poll::Ready` will result in it attempting to acquire another permit. If
the semaphore only has a single permit, and the gate has already
acquired it, it will never become ready if `poll_ready` is called a
second time.

This commit fixes this by only calling `self.poll_acquire()` if the gate
has not acquired a permit. If it _has_ acquired a permit, the permit is
temporarily moved back out of the gate while polling the inner service,
in case that service has become unready. This way, the gate will not
attempt to acquire a second permit when it already has one, and will
release the permit if the inner service is no longer ready.