outbound: Configure balancers with service metadata
The OutboundPolicies API includes resource references with its
responses. These references allow us to accurately identify an arbitrary
(Kubernetes style) resource by its group, kind, namespace, and name.

To support new metrics that include these resource coordinates as
labels, we need access to the parent and backend references from the
balancer stack.

This change makes the following changes to accomodate this:

1. We introduce `ParentRef`, `BackendRef`, and `EndpointRef` new-types
   to serve as explicit markers for different types of metadata we may
   encounter.
2. The profile stack is updated to parse out metadata from service names
   in the form <name>.<ns>.svc.*.
3. We also support discovering pod metadata from the endpoint labels
   `dst_pod` and `dst_namespace`.
4. When we can't infer any metadata from a profile response, we use an
   "unknown" metadata.
5. When using policy, we use control-plane provided metadata.

In practice, pod metadata won't be surfaced by the existing code path;
but we have a relatively simple path forward to using it.

This change splits the balancer stack into its own layer so it's
distinct from the rest of the concrete stack configuration.

There are only minor functional changes in this branch:

1. The `balance{addr=...}` tracing context is replace by
   `service{ns=..,name=...}` for clarity;
2. Errors returned from the balance stack are now rendered as:
   "Service <name>.<ns>: <error>"