outbound: add `l5d-dst-canonical` to requests with ServiceProfiles (#2383)
PR #2250 removed the `l5d-dst-canonical` header from ServiceProfile
requests. This header is used by the inbound proxy for telemetry
purposes, so removing it and not putting it back broke ServiceProfile
route metrics. This commit adds a layer for setting this header to the
service profile route stack.

We can also add the header for non-ServiceProfile requests if that's
desirable...I'll have to look into whether it is. This commit should, at
least, fix the existing ServiceProfile route metrics.

Fixes #10521