outbound: Report HTTP balancer endpoint gauges (#2380)
It's not currently possible to know how many endpoints are in a
balancer. This change adds an `outbound_http_balancer_endpoints` gauge
that exposes the number of endpoints a balancer has by their current
readiness status.

Note that in the concrete stack we do not currently differentiate
between gRPC and HTTP backends, so all balancers are exposed under this
single metric.