add test for inbound opaque ports not in default ports
This adds a test for the handling of inbound opaque connections when the
port is *not* in the `LINKERD2_PROXY_INBOUND_PORTS` environment variable
and the opaque policy is not eagerly discovered from the control plane
on startup. Currently, this test fails because the proxy does not know
to handle connections on this port as opaque, and attempts to perform
protocol detection instead.