inbound: add opaque port ranges to `policy::Store`
This changes the `policy::Store` type to accept a set of port ranges
which should be marked as opaque. This allows the default policy used to
initialize the policy watch to differ based on whether or not a port is
marked as opaque. This way, when starting discovery for an opaque port,
an opaque default policy will be used until the port's actual policy is
discovered from the control plane.