Classify grpc requests properly in the http classifier (#2410)
The gRPC protocol always sets the HTTP response status code to 200 and instead communicates failures in a grpc-status header sent in a TRAILERS frame. Linkerd uses the HTTP response status code to determine if a response is successful, and therefore will consider all gRPC responses successful regardless of their gRPC status code. This means that functionality such as retries and circuit breaking do not function correctly with gRPC traffic.

We update the Http classifier to look for the presence of a `Content-Type: application/grpc` header and use Grpc response classification when it is set.

Signed-off-by: Alex Leong <alex@buoyant.io>